Specialization insert, update and delete fixed #89 #91 #93

+++ dddnetcore/Domain/Specializations/SpecializationService.cs
@@ -1,6 +1,13 @@
 using System.Threading.Tasks;
 using System.Collections.Generic;
 using DDDSample1.Domain.Shared;
+using DDDSample1.Domain.Staffs;
+using System.Formats.Asn1;
+using Microsoft.CodeAnalysis;
+using DDDSample1.Domain.OperationTypes;
+using System;
+using DDDSample1.Domain.OperationTypesSpecializations;
+using System.Linq;
 
 namespace DDDSample1.Domain.Specializations
 {
@@ -8,11 +15,15 @@ namespace DDDSample1.Domain.Specializations
     {
         private readonly IUnitOfWork _unitOfWork;
         private readonly ISpecializationRepository _repo;
+        private readonly IStaffRepository _staffRepo;
+        private readonly IOperationTypeSpecializationRepository _operationTypeRepo;
 
-        public SpecializationService(IUnitOfWork unitOfWork, ISpecializationRepository repo)
+        public SpecializationService(IUnitOfWork unitOfWork, ISpecializationRepository repo, IOperationTypeSpecializationRepository operationTypeRepository, IStaffRepository staffRepository)
         {
             this._unitOfWork = unitOfWork;
             this._repo = repo;
+            this._staffRepo=staffRepository;
+            this._operationTypeRepo=operationTypeRepository;
         }
 
         public async Task<List<SpecializationDto>> GetAllAsync()
@@ -23,7 +34,9 @@ namespace DDDSample1.Domain.Specializations
                 new SpecializationDto
                 {
                     Id = spec.Id.AsGuid(),
-                    SpecializationName = spec.SpecializationName.ToString()
+                    SpecializationName = spec.SpecializationName.ToString(),
+                    SpecializationCode = spec.SpecializationCode.ToString(),
+                    SpecializationDescription = spec.SpecializationDescription.ToString(),
                 });
 
             return listDto;
@@ -46,7 +59,9 @@ namespace DDDSample1.Domain.Specializations
         public async Task<SpecializationDto> AddAsync(CreatingSpecializationDto dto)
         {
             var specialization = new Specialization(
-                new SpecializationName(dto.SpecializationName)
+                new SpecializationName(dto.SpecializationName),
+                new SpecializationCode(dto.SpecializationCode),
+                new SpecializationDescription(dto.SpecializationDescription)
             );
 
             await this._repo.AddAsync(specialization);
@@ -55,7 +70,9 @@ namespace DDDSample1.Domain.Specializations
             return new SpecializationDto
             {
                 Id = specialization.Id.AsGuid(),
-                SpecializationName = specialization.SpecializationName.ToString()
+                SpecializationName = specialization.SpecializationName.ToString(),
+                SpecializationCode = specialization.SpecializationCode.ToString(),
+                SpecializationDescription = specialization.SpecializationDescription.ToString(),
             };
         }
 
@@ -67,13 +84,18 @@ namespace DDDSample1.Domain.Specializations
                 return null;
 
             specialization.ChangeSpecializationName(new SpecializationName(dto.SpecializationName));
+            specialization.ChangeSpecializationCode(new SpecializationCode(dto.SpecializationCode));
+            specialization.UpdateSpecializationDescription(new SpecializationDescription(dto.SpecializationDescription));
             
             await this._unitOfWork.CommitAsync();
 
             return new SpecializationDto
             {
                 Id = specialization.Id.AsGuid(),
-                SpecializationName = specialization.SpecializationName.ToString()
+                SpecializationName = specialization.SpecializationName.ToString(),
+                SpecializationCode = specialization.SpecializationCode.ToString(),
+                SpecializationDescription = specialization.SpecializationDescription.ToString(),
+
             };
         }
 
@@ -89,7 +111,9 @@ namespace DDDSample1.Domain.Specializations
             return new SpecializationDto
             {
                 Id = specialization.Id.AsGuid(),
-                SpecializationName = specialization.SpecializationName.ToString()
+                SpecializationName = specialization.SpecializationName.ToString(),
+                SpecializationCode = specialization.SpecializationCode.ToString(),
+                SpecializationDescription = specialization.SpecializationDescription.ToString(),
             };
         }
 
@@ -100,14 +124,45 @@ namespace DDDSample1.Domain.Specializations
             if (specialization == null)
                 return null;
             
+            if(await CheckSpecializationIsAtributtedToStaff(specialization))
+            {
+                throw new BusinessRuleValidationException("Specialization is atributted to staff, can´t delete it");
+            }
+
+            if(await CheckSpecializationIsAtributtedToOpType(specialization))
+            {
+                throw new BusinessRuleValidationException("Specialization is atributted to opType, can´t delete it");
+            }
+
+
+            
             this._repo.Remove(specialization);
             await this._unitOfWork.CommitAsync();
 
             return new SpecializationDto
             {
                 Id = specialization.Id.AsGuid(),
-                SpecializationName = specialization.SpecializationName.ToString()
+                SpecializationName = specialization.SpecializationName.ToString(),
+                SpecializationCode = specialization.SpecializationCode.ToString(),
+                SpecializationDescription = specialization.SpecializationDescription.ToString(),
             };
+        }
+
+        public async Task<Boolean> CheckSpecializationIsAtributtedToStaff(Specialization specialization){
+            var list= await _staffRepo.CheckSpecializationIsAtributtedToStaff(specialization);
+            Console.WriteLine(list.Any());
+            if(list.Any()){
+                return true;
+            }
+            return false;
+        }
+
+        public async Task<Boolean> CheckSpecializationIsAtributtedToOpType(Specialization specialization){
+            var list = await _operationTypeRepo.CheckSpecializationIsAtributtedToOpType(specialization);
+            if(list.Any()){
+                return true;
+            }
+            return false;
         }
     }
 }
